Conditional logic in before-install.sh

+++ before-install.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 
-sudo rm /usr/share/arete-lib.sh 2>/dev/null
-sudo rm /usr/share/arete.sh 2>/dev/null
-sudo unlink /usr/sbin/arete 2>/dev/null
+if [ -f /usr/share/arete-lib.sh ]; then
+    sudo rm /usr/share/arete-lib.sh 2>/dev/null
+fi
+
+if [ -f /usr/share/arete.sh ]; then
+    sudo rm /usr/share/arete.sh 2>/dev/null
+fi
+
+if -f [ /usr/sbin/arete ]; then
+    sudo unlink /usr/sbin/arete 2>/dev/null
+fi
